+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/OidcUserInfo.kt
@@ -8,6 +8,7 @@ import kotlinx.serialization.Serializable
 /**
  * [OpenID Connect Core 1.0](https://openid.net/specs/openid-connect-core-1_0.html)
  */
+@Serializable
 data class OidcUserInfo(
     @SerialName("sub")
     val subject: String,